refactor folder mapping and comment-ing home
