+++ pages/home/index.module.css
@@ -1,0 +1,25 @@
+.section {
+    width: 60%;
+    margin: auto;
+    margin-bottom: 1rem;
+    margin-top: 1.5rem;
+}
+
+.sectionHead {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 0.5rem;
+}
+
+.a {
+    color: 	#4CB654;
+    text-decoration: none;
+}
+
+.sectionBody {
+    display: flex;
+    margin: 0px 4px;
+    overflow: auto;
+    scroll-behavior: smooth;
+}
