+++ pages/show-redux-state/index.module.css
@@ -1,0 +1,7 @@
+.codeStyle {
+    background: '#ebebeb';
+    width: 400;
+    padding: 10;
+    border: '1px solid grey';
+    margin-bottom: 10;
+  }
